11408 - moved primary to it's own component
